% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationList.R
\name{classificationList}
\alias{classificationList}
\title{Retrieve lists of classification tables from the CELLAR and FAO repositories}
\usage{
classificationList(endpoint = "ALL", showQuery = FALSE)
}
\arguments{
\item{endpoint}{One of "CELLAR", "FAO", or "ALL" (default).}

\item{showQuery}{Logical. If TRUE, returns the SPARQL query along with the data.}
}
\value{
A data frame (if endpoint is "CELLAR" or "FAO"), or a named list of two data frames (if endpoint is "ALL").
}
\description{
Retrieve a list of classification tables from the CELLAR and FAO repositories.
}
\details{
The behaviour of this function is contingent on the global option \code{useLocalDataForVignettes}:
The default behaviour (when the option is not set, or set to something else than \code{TRUE}), it queries live SPARQL endpoints online.
When the option is set to \code{TRUE} via \code{options(useLocalDataForVignettes = TRUE)}, the function returns local (embedded) data instead of querying live SPARQL endpoints.
This is useful for building vignettes or offline testing.
}
