% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixList.R
\name{prefixList}
\alias{prefixList}
\title{Create a list of prefixes for both CELLAR and FAO repositories.}
\usage{
prefixList(endpoint, prefix = NULL)
}
\arguments{
\item{endpoint}{SPARQL endpoint to query. Must be either \code{"CELLAR"} or \code{"FAO"}.}

\item{prefix}{Optional. A vector of prefix names to filter. If \code{NULL}, all available prefixes will be returned.}
}
\value{
A character matrix of SPARQL PREFIX declarations.
}
\description{
Creates a list of prefixes to be used when building SPARQL queries to retrieve classification tables from CELLAR or FAO.
If the global option \code{useLocalDataForVignettes} is set to TRUE (e.g. \code{options(useLocalDataForVignettes = TRUE)}),
the function uses local pre-saved metadata instead of querying live endpoints.
}
\details{
The behaviour of this function is contingent on the global option \code{useLocalDataForVignettes}:
The default behaviour (when the option is not set, or set to something else than \code{TRUE}), it queries live SPARQL endpoints online.
When the option is set to \code{TRUE} via \code{options(useLocalDataForVignettes = TRUE)}, the function returns local (embedded) data instead of querying live SPARQL endpoints.
This is useful for building vignettes or offline testing.
}
